server/sonar-db-migration/src/main/java/org/sonar/server/platform/db/migration/version/v99/UpdateUserUuidColumnSizeInAuditTable.java
/*
 * SonarQube
 * Copyright (C) 2009-2024 SonarSource SA
 * mailto:info AT sonarsource DOT com
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3 of the License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public License
 * along with this program; if not, write to the Free Software Foundation,
 * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
package org.sonar.server.platform.db.migration.version.v99;

import com.google.common.annotations.VisibleForTesting;
import java.sql.SQLException;
import org.sonar.db.Database;
import org.sonar.server.platform.db.migration.def.VarcharColumnDef;
import org.sonar.server.platform.db.migration.sql.AlterColumnsBuilder;
import org.sonar.server.platform.db.migration.step.DdlChange;

import static org.sonar.server.platform.db.migration.def.VarcharColumnDef.newVarcharColumnDefBuilder;

public class UpdateUserUuidColumnSizeInAuditTable extends DdlChange {

  @VisibleForTesting
  static final String TABLE_NAME = "audits";

  @VisibleForTesting
  static final String COLUMN_NAME = "user_uuid";

  private static final VarcharColumnDef columnDefinition = newVarcharColumnDefBuilder()
    .setColumnName(COLUMN_NAME)
    .setIsNullable(false)
    .setLimit(255)
    .build();

  public UpdateUserUuidColumnSizeInAuditTable(Database db) {
    super(db);
  }

  @Override
  public void execute(Context context) throws SQLException {
    context.execute(new AlterColumnsBuilder(getDialect(), TABLE_NAME)
      .updateColumn(columnDefinition)
      .build());


  }
}
